✨ feat(校验tag的删除、优化代码):

+++ server/src/routes/tag/delete-tag.ts
@@ -1,14 +1,35 @@
 import Router from "@koa/router";
 import DB from "@/db";
 import sequelize from "@/db/config";
-import { Op } from "sequelize";
 import interger from "@/common/verify/integer";
-import authMiddleware from "@/common/middleware/auth";
+import auth from "@/common/middleware/auth";
+import stringArrayReplace from "@/db/utils/stringArrayReplace";
 
 let router = new Router();
 
-router.delete("/tag/:id", interger([], ["id"]),authMiddleware(), async ctx => {
+router.delete("/tag/:id", interger([], ["id"]), auth(), async ctx => {
   let { id } = ctx.params;
+
+  //判断是否有文章只有这一个tag
+  let allowDelete = await DB.Article.count({ attributes: ["id"], where: { tag: id } })
+    .then(count => ({
+      success: !count,
+      mes: `有${count}篇文章只使用了这个Tag，需要在删除前进行处理。`,
+    }))
+    .catch(err => {
+      console.log(err);
+      return { success: false, mes: "检测使用Tag的文章数量时出错" };
+    });
+
+  if (!allowDelete.success) {
+    ctx.status = 409; //请求与服务器的当前状态冲突，无法完成请求
+    ctx.body = {
+      success: false,
+      message: allowDelete.mes,
+    };
+    return;
+  }
+
   const t = await sequelize.transaction();
   let deleteTagCount = await DB.Tag.destroy({
     where: {
@@ -17,38 +38,20 @@ router.delete("/tag/:id", interger([], ["id"]),authMiddleware(), async ctx => {
     transaction: t,
   });
 
-  /** 需要删除的个数*/
-  let { count: tagCount } = await DB.Article.findAndCountAll({
-    where: { tag: { [Op.substring]: id } },
-  });
-  //将文章表中的 ,id,  ,id   id,   id  四种方法全部置换
-  let replaceTagIdResult = await Promise.all<any>([
-    sequelize.query(`update article set tag=REPLACE (tag,',${id},',',') WHERE tag like '%${id}%'`, {
+  let replaceTagIdResult = await stringArrayReplace(
+    {
+      tableName: "Article",
+      field: "tag",
+      oldValue: id,
+      newValue: "",
+    },
+    {
       transaction: t,
-    }),
-    sequelize.query(`update article set tag=REPLACE (tag,',${id}','') WHERE tag like '%${id}%'`, {
-      transaction: t,
-    }),
-    sequelize.query(`update article set tag=REPLACE (tag,'${id},','') WHERE tag like '%${id}%'`, {
-      transaction: t,
-    }),
-    sequelize.query(`update article set tag=REPLACE (tag,'${id}','') WHERE tag like '%${id}%'`, {
-      transaction: t,
-    }),
-  ]).then(res => {
-    // 实际删除了多少
-    let _tagCount = res.reduce((total, item) => {
-      return total + item[0].affectedRows;
-    }, 0);
-    if (_tagCount != tagCount) {
-      console.log(`删除 ${id} 时 置换数量对不上${_tagCount}/${tagCount}`);
-      return false;
-    } else {
-      return true;
     }
-  });
+  );
 
   if (!(deleteTagCount && replaceTagIdResult)) {
+    ctx.status = 500;
     ctx.body = {
       success: false,
       message: "删除失败",
